--- site/src/App.styl
@@ -1,47 +1,0 @@
-@import "./styles/settings/_variables"
-@import "./styles/helpers/_functions"
-@import "./styles/partials/index"
-
-/*!
-*
-* Box model reset
-* http://css-tricks.com/inheriting-box-sizing-probably-slightly-better-best-practice/
-*
-*/
-
-html
-  font-size: 100%
-  box-sizing: border-box
-
-*, *:before, *:after
-  box-sizing: inherit
-
-body
-  font: normal normal  $base-font-size\/em($base-line-height) $font-family-base
-  background-color: darken($bg-color, 30%)
-  color: $base-text-color
-  text-rendering: optimizeLegibility
-  -webkit-font-smoothing: antialiased
-  -moz-osx-font-smoothing: grayscale
-  font-smoothing: auto
-  font-feature-settings: "liga", "kern"
-
-::selection
-  background: $selection-bg-color
-  color: $selection-color
-
-// /*!
-// *
-// * Better Font Smoothing in Chrome & Safari
-// * http://www.intridea.com/blog/2014/5/8/better-font-smoothing-in-chrome-on-mac-os-x
-// *
-// */
-@media only screen and (-webkit-min-device-pixel-ratio: 1.25),\
-  only screen and ( min-device-pixel-ratio: 1.25),\
-  only screen and ( min-resolution: 200dpi),\
-  only screen and ( min-resolution: 1.25dppx)
-
-  body
-    -webkit-font-smoothing: subpixel-antialiased
-
-
