+++ site/src/Toast.module.scss
@@ -1,0 +1,7 @@
+.toast {
+  justify-content: center;
+  background-color: #13323d;
+  box-shadow: -20px 20px 80px 20px rgba(#13323d, 0.2);
+  color: white;
+  font-weight: bold;
+}
