--- site/src/Toast.module.styl
@@ -1,10 +1,0 @@
-.toast
-  .toast-content
-    font-weight: bold
-    text-align: center
-    color: white
-    background-color: #13323D
-    box-shadow: -20px 20px 80px 20px rgba(#13323D,0.2)
-
-    button
-      display: none
