+++ site/src/components/Footer/Footer.module.scss
@@ -1,0 +1,5 @@
+.footer {
+  overflow: hidden;
+  padding: 0 20px;
+  text-align: center;
+}
