--- site/src/components/Footer/Footer.module.styl
@@ -1,4 +1,0 @@
-.footer
-  overflow: hidden
-  text-align: center
-  padding: 0 20px
