+++ site/src/components/Footer/Footer.tsx
@@ -1,4 +1,4 @@
-import styles from './Footer.module.styl'
+import styles from './Footer.module.scss'
 
 export const Footer = () => {
   const date = new Date()
