+++ site/src/components/Header/Header.module.scss
@@ -1,0 +1,59 @@
+@use 'sass:color';
+@use '../../styles/settings/variables' as v;
+
+.panel {
+  position: relative;
+  z-index: 1;
+  padding: v.$base-line-height 20px;
+  background: color.adjust(v.$bg-color, $lightness: -10%) 0%;
+  background: linear-gradient(
+    to bottom,
+    color.adjust(v.$bg-color, $lightness: -10%) 0%,
+    color.adjust(v.$bg-color, $lightness: -20%) 100%
+  );
+  box-shadow: 0 2px 6px color.adjust(#000, $lightness: 20%);
+  text-align: center;
+
+  abbr {
+    text-decoration: none;
+  }
+
+  span {
+    @media (max-width: v.$phone) {
+      display: block;
+    }
+  }
+
+  &__title {
+    margin-top: 0;
+  }
+
+  .logo {
+    display: block;
+    max-height: v.$base-line-height * 4;
+    margin: 0 auto v.$base-line-height;
+  }
+
+  .copy {
+    border-radius: 3px;
+    margin-bottom: 1.5em;
+    cursor: pointer;
+
+    code {
+      display: inline-block !important;
+
+      @media (min-width: v.$phone) {
+        padding-right: 40px !important;
+      }
+
+      &:hover {
+        @media (min-width: v.$phone) {
+          background-image: url('./copy.svg');
+          background-position: center right 10px;
+          background-repeat: no-repeat;
+          background-size: 20px 20px;
+        }
+      }
+    }
+  }
+}
