--- site/src/components/Header/Header.module.styl
@@ -1,49 +1,0 @@
-@import '../../styles/settings/_variables';
-@import '../../styles/helpers/_vertical-rhythm'
-
-.panel
-  background: linear-gradient(
-    to bottom, darken($bg-color, 30%) 0%,
-    darken($bg-color, 45%) 100%
-  );
-  padding: $base-line-height 20px
-  text-align: center
-  box-shadow: 0 2px 6px lighten(#000, 20%)
-  position relative
-  z-index: 1
-
-  abbr
-    text-decoration: none
-
-  span
-    @media (max-width: $phone)
-      display: block
-
-  &__title
-    font-family: 'PT Sans', sans-serif
-    font-weight: 700
-    margin-top: 0
-
-.logo
-  max-height: $base-line-height * 4
-  display: block
-  margin: 0 auto $base-line-height
-
-.copy
-  cursor: pointer
-  border-radius: 3px
-  trailer(1)
-
-  code
-    display: inline-block !important
-
-    @media (min-width: $phone)
-      padding-right: 40px !important
-
-    &:hover
-      @media (min-width: $phone)
-        background-position: center right 10px
-        background-repeat: no-repeat
-        background-size: 20px 20px
-        background-image: url('./copy.svg')
-
