+++ site/src/components/Header/components/ChangeLocale/ChangeLocale.module.scss
@@ -1,0 +1,21 @@
+@use '../../../../styles/settings/_variables' as v;
+
+.box {
+  text-align: left;
+
+  @media (min-width: v.$tablet) {
+    position: absolute;
+  }
+}
+
+.button {
+  border: none;
+  background: none;
+  color: white;
+  cursor: pointer;
+  outline: none;
+}
+
+.selected {
+  text-decoration: underline;
+}
