--- site/src/components/Header/components/ChangeLocale/ChangeLocale.module.styl
@@ -1,17 +1,0 @@
-@import '../../../../styles/settings/_variables'
-@import '../../../../styles/helpers/_vertical-rhythm'
-
-.box
-  text-align: left
-  @media (min-width: $tablet)
-    position absolute
-
-.button
-  border: none
-  background: none
-  color: white
-  cursor: pointer
-  outline: none
-
-.selected
-  text-decoration: underline
