+++ site/src/components/Header/components/Donate/Code.module.scss
@@ -1,0 +1,96 @@
+.code-box {
+  position: relative;
+  z-index: 1;
+  top: 0;
+  left: 0;
+  display: none;
+  width: 100%;
+  height: 100%;
+  background-color: rgb(255 255 255 / 30%);
+  perspective: 400px;
+}
+
+.code {
+  position: absolute;
+  top: calc(50% - 100px);
+  left: calc(50% - 100px);
+  overflow: hidden;
+  width: 200px;
+  height: 200px;
+  border-radius: 6px;
+  background: #fff no-repeat center center;
+  background-size: 190px;
+  box-shadow: 0 2px 7px rgb(0 0 0 / 30%);
+  cursor: pointer;
+  opacity: 0;
+  text-align: center;
+  transform-origin: center center;
+  transform-style: preserve-3d;
+  transition: all 1s ease-in-out;
+}
+
+.show-code {
+  animation: show-qr 3s ease-in-out;
+  animation-fill-mode: forwards;
+  animation-iteration-count: 1;
+  opacity: 1;
+}
+
+.hide-code {
+  animation: hide-qr 0.5s ease-in-out 0;
+  animation-fill-mode: forwards;
+  animation-iteration-count: 1;
+  opacity: 1;
+}
+
+@keyframes show-qr {
+  0% {
+    transform: rotateX(90deg);
+  }
+
+  8% {
+    opacity: 1;
+    transform: rotateX(-60deg);
+  }
+
+  18% {
+    transform: rotateX(40deg);
+  }
+
+  34% {
+    transform: rotateX(-28deg);
+  }
+
+  44% {
+    transform: rotateX(18deg);
+  }
+
+  58% {
+    transform: rotateX(-12deg);
+  }
+
+  72% {
+    transform: rotateX(9deg);
+  }
+
+  88% {
+    transform: rotateX(-5deg);
+  }
+
+  96% {
+    transform: rotateX(2deg);
+  }
+}
+
+@keyframes hide-qr {
+  20%,
+  50% {
+    opacity: 1;
+    transform: scale(1.08, 1.08);
+  }
+
+  100% {
+    opacity: 0;
+    transform: rotateZ(40deg) scale(0.6, 0.6);
+  }
+}
