--- site/src/components/Header/components/Donate/Code.module.styl
@@ -1,83 +1,0 @@
-
-.code-box
-  position: relative
-  width: 100%
-  height: 100%
-  top: 0
-  left: 0
-  background-color: rgba(255,255,255,0.3)
-  perspective: 400px
-  display: none
-  z-index: 1
-
-.code
-  position: absolute
-  cursor: pointer
-  text-align: center
-  width: 200px
-  height: 200px
-  left: calc(50% - 100px)
-  top: calc(50% - 100px)
-  background: #fff no-repeat center center
-  background-size: 190px
-  border-radius: 6px
-  box-shadow: 0px 2px 7px rgba(0,0,0,0.3)
-  opacity: 0
-  transition: all 1s ease-in-out
-  transform-style: preserve-3d
-  transform-origin: center center
-  overflow: hidden
-
-.show-code
-  opacity: 1
-  animation: showQR 3s ease-in-out
-  animation-iteration-count: 1
-  animation-fill-mode: forwards
-
-.hide-code
-  opacity: 1
-  animation: hideQR 0.5s ease-in-out 0
-  animation-iteration-count: 1
-  animation-fill-mode: forwards
-
-@keyframes showQR {
-  from {
-    transform: rotateX(90deg);
-  }
-  8% {
-    opacity: 1;
-    transform: rotateX(-60deg);
-  }
-  18% {
-    transform: rotateX(40deg);
-  }
-  34% {
-    transform: rotateX(-28deg);
-  }
-  44% {
-    transform: rotateX(18deg);
-  }
-  58% {
-    transform: rotateX(-12deg);
-  }
-  72% {
-    transform: rotateX(9deg);
-  }
-  88% {
-    transform: rotateX(-5deg);
-  }
-  96% {
-    transform: rotateX(2deg);
-  }
-}
-
-@keyframes hideQR {
-  20%, 50% {
-    transform: scale(1.08,1.08);
-    opacity: 1;
-  }
-  to {
-    opacity: 0;
-    transform: rotateZ(40deg) scale(0.6,0.6);
-  }
-}
