--- site/src/components/Header/components/Donate/Donation.module.styl
@@ -1,107 +1,0 @@
-.list
-    list-style: none
-    list-style-type: none
-    display: table
-    margin: auto
-    padding: 0
-    transition: all .3s
-    z-index: 1
-    position relative
-
-    li
-      float: left
-      border-left: 1px solid #ddd
-      border: 1px  solid #ddd
-      background-color: white
-
-      &:first-child
-        border-radius: 6px 0 0 6px
-        border-right: 0
-
-      &:last-child
-        border-radius: 0 6px 6px 0
-        border-left: 0
-
-.blur
-  filter: blur(3px)
-
-
-.button
-  display: block
-  width: 74px
-  height: 26px
-  background: no-repeat center center
-  background-size: 45px
-  color: black
-  text-decoration: none
-  text-align: center
-  line-height: 999
-  border-radius: 0
-  padding: 0
-  transition: all .3s
-  filter: grayscale(1)
-  opacity: .6
-  border: none
-  cursor: pointer
-
-  &:visited
-    color: black
-
-  &:hover
-    background-color: rgba(204, 217, 220, 0.5)
-    filter: grayscale(0)
-    opacity: 1
-
-  &--star
-    line-height: 28px
-
-
-  &--paypal
-    background-image: url('./img/paypal.svg')
-
-  &--bitcoin
-    background-image: url('./img/bitcoin.svg')
-
-  &--litcoin
-    background-image: url('./img/litecoin.svg')
-
-
-
-[data-footnote]
-  position: relative
-  overflow: hidden
-
-  &:hover
-    overflow: visible
-
-    &:before,
-    &:after
-      opacity: 1
-      transform: translate3d(-50%,-7px,0)
-
-  &:before,
-  &:after
-    position: absolute
-    transition: all .3s
-    transform: translate3d(-50%,0,0)
-    opacity: 0
-    left: 37px
-    z-index: 10
-
-  &:before
-    content: attr(data-footnote)
-    border-radius: 4px
-    background-color: lighten(#000,17%)
-    color: #fff
-    height: 26px
-    line-height: 26px
-    padding: 0 6px
-    font-size: 12px
-    white-space: nowrap
-    top: -24px
-
-  &:after
-    content: ''
-    border: 5px solid lighten(#000,17%)
-    border-color: lighten(#000,17%) transparent transparent transparent
-    top: 0
