+++ site/src/components/Main/components/GenerateSection/GenerateSection.module.scss
@@ -1,0 +1,52 @@
+@use 'sass:color';
+@use '../../../../styles/settings/variables' as v;
+
+.generate-section {
+  overflow: auto;
+  background: #fff;
+  color: v.$bg-color;
+  text-align: center;
+  user-select: none;
+}
+
+.generate-button {
+  display: inline-block;
+  overflow: hidden;
+  padding: 0.375em 25px;
+  border: none;
+  border-radius: 0 6px 6px 0;
+  margin-bottom: 0.75em;
+  background-color: color.adjust(v.$bg-color, $lightness: -60%);
+  box-shadow:
+    0 1px 2px rgb(0 0 0 / 10%),
+    0 1px 2px rgb(0 0 0 / 10%);
+  color: #fff;
+  cursor: pointer;
+  font-size: 1em;
+  line-height: 1.5em;
+  outline: none;
+  text-align: center;
+  text-decoration: none;
+  text-transform: lowercase;
+  transition: 0.15s ease-out;
+  user-select: none;
+  vertical-align: bottom;
+
+  @media (max-width: v.$phone) {
+    width: 100%;
+    border-radius: 6px;
+  }
+}
+
+.input {
+  border-radius: 6px 0 0 6px;
+  background: #ededed;
+  cursor: pointer;
+  text-align: center;
+  user-select: none;
+
+  @media (max-width: v.$phone) {
+    width: 100%;
+    border-radius: 6px;
+  }
+}
