--- site/src/components/Main/components/GenerateSection/GenerateSection.module.styl
@@ -1,52 +1,0 @@
-@import '../../../../styles/helpers/_vertical-rhythm'
-@import '../../../../styles/settings/_variables'
-@import '../_center'
-
-.generate-section
-  user-select: none
-  text-align: center
-  background: #FFF
-  color: $bg-color
-
-.center
-  @extend $center
-
-.generate-button
-  user-select: none
-  border: none
-  border-radius: 0 6px 6px 0
-  display: inline-block
-  outline: none
-  overflow: hidden
-  color: #FFF
-  text-align: center
-  text-decoration: none
-  vertical-align: bottom
-  box-shadow: 0 1px 2px rgba(0, 0, 0, .1), 0 1px 2px rgba(0, 0, 0, 0.1)
-  transition: .15s ease-out
-  cursor: pointer
-  text-transform: lowercase
-  padding: 25px
-  rhythm-padding(.25)
-  adjust-font-size-to($base-font-size)
-  trailer(.5)
-  background-color: darken($bg-color, 60%)
-
-  &:hover:not([disabled])
-    background-color: shade($bg-color, 20%)
-
-  @media (max-width: $phone)
-    border-radius: 6px
-    width: 100%
-
-
-.input
-  user-select: none
-  border-radius: 6px 0 0 6px
-  cursor: pointer
-  text-align: center
-  background: #ededed
-
-  @media (max-width: $phone)
-    border-radius: 6px
-    width: 100%
