--- site/src/components/Main/components/InfoSection/InfoSection.module.styl
@@ -1,12 +1,0 @@
-@import '../../../../styles/settings/_variables'
-@import '../_center'
-
-.info-section
-  background: $bg-color
-
-.center
-  @extend $center
-
-.list
-  padding-left: 20px
-  list-style-type: none
