--- site/src/components/Main/components/ValidateSection/ValidateSection.module.styl
@@ -1,43 +1,0 @@
-@import '../../../../styles/helpers/_vertical-rhythm'
-@import '../../../../styles/settings/_variables'
-@import '../_center'
-
-.validate-section
-  user-select: none
-  text-align: center
-  background: #FFF
-  color: $bg-color
-
-  &__input
-    border-radius: 6px 0 0 6px
-    padding: 10px
-    text-align: center
-    background: #ededed
-    line-height: 1
-
-
-    @media (max-width: $phone)
-      border-radius: 6px !important
-      width: 100%
-
-
-
-.center
-  padding-trailer(2)
-  @extend $center
-
-.message
-  border-radius: 0 6px 6px 0
-  color: lighten(#000, 25%)
-  background-color: darken(#FFF, 28%)
-  transition: all ease 1s
-  display: inline-block
-
-  &--valid
-    color: $color = green
-    background-color: lighten($color, 70%)
-
-  &--invalid
-    color: $color = red
-    background-color: lighten($color, 70%)
-
