--- site/src/components/Main/components/_center.styl
@@ -1,5 +1,0 @@
-$center
-  max-width: 1000px
-  padding: 0 20px
-  margin: auto
-  overflow: auto
