+++ site/src/styles/helpers/_functions.scss
@@ -1,0 +1,33 @@
+@use 'sass:math';
+@use 'sass:meta';
+@use '../settings/variables' as v;
+
+@function clean-unit($value) {
+  @if meta.type-of($value) == 'number' {
+    @return math.div($value, $value * 0 + 1);
+  } @else {
+    @warn "#{$value} it's not a number";
+  }
+}
+
+// EM function
+// Convert pixel units to em units
+
+@function em($pixels, $context: v.$base-font-size) {
+  @if meta.type-of($pixels) == 'number' {
+    @return (math.div(clean-unit($pixels), clean-unit($context))) * 1em;
+  } @else {
+    @warn '#{$pixels} is not a number';
+  }
+}
+
+// REM function
+// Convert pixel units to rem units
+
+@function rem($pixels, $context: v.$base-font-size) {
+  @if meta.type-of($pixels) == 'number' {
+    @return (math.div(clean-unit($pixels), clean-unit($base-font-size))) * 1rem;
+  } @else {
+    @warn '#{$pixels} is not a number';
+  }
+}
