--- site/src/styles/helpers/_functions.styl
@@ -1,11 +1,0 @@
-/* EM function */
-// Convert pixel units to em units
-
-em($pixels, $context = $base-font-size)
-    (floor($pixels/$context, 10))em
-
-/* REM function */
-// Convert pixel units to rem units
-
-rem($pixels)
-    (floor($pixels/$base-font-size, 10))rem
