--- site/src/styles/partials/blockquote.styl
@@ -1,8 +1,0 @@
-@import '../settings/_variables'
-@import '../helpers/_vertical-rhythm'
-
-blockquote
-  font-style: italic
-  padding-left: em(10px)
-  border-left: em(5px) solid lighten(#000, 60%)
-  rhythm-margins(1, 1, $base-font-size)
