+++ site/src/styles/partials/code.scss
@@ -1,0 +1,9 @@
+@use '../settings/_variables' as v;
+
+code {
+  display: inline-block;
+  overflow: auto;
+  padding: 0 #{v.$base-line-height} / 2;
+  border-radius: 3px;
+  vertical-align: bottom;
+}
