--- site/src/styles/partials/code.styl
@@ -1,9 +1,0 @@
-@import '../settings/_variables'
-@import '../helpers/_vertical-rhythm'
-
-code
-  padding: 0 0.2em
-  display: inline-block
-  border-radius: 3px
-  vertical-align: bottom
-  overflow: auto
