+++ site/src/styles/partials/headers.scss
@@ -1,0 +1,22 @@
+%font-family {
+  font-family: 'PT Sans', sans-serif;
+}
+
+h1 {
+  @extend %font-family;
+
+  margin-bottom: 0.6666em;
+  font-size: 2.25em;
+  font-weight: 700;
+  line-height: 1.3333em;
+}
+
+h2 {
+  @extend %font-family;
+
+  margin-top: 0.75em;
+  margin-bottom: 0.75em;
+  font-size: 2em;
+  font-weight: 400;
+  line-height: 1.5em;
+}
