--- site/src/styles/partials/headers.styl
@@ -1,8 +1,0 @@
-h1, h2, h3, h4, h5, h6
-  font-family: 'PT Sans', sans-serif
-  font-weight: 400
-
-for num in (1..6)
-  h{num}
-      adjust-font-size-to($font-size-h + num)
-      rhythm-margins(1, 1, $font-size-h + num)
