--- site/src/styles/partials/index.styl
@@ -1,11 +1,0 @@
-@import '../settings/_variables';
-@import '../helpers/_functions';
-@import '../helpers/_vertical-rhythm';
-
-@import 'blockquote';
-@import 'code';
-@import 'headers';
-@import 'input';
-@import 'links';
-@import 'paragraph';
-@import 'pre';
