--- site/src/styles/partials/input.styl
@@ -1,32 +1,0 @@
-@import '../helpers/_functions'
-
-$input
-  adjust-font-size-to($base-font-size)
-  border: none
-  color: lighten(#000, 35%)
-  trailer(.5)
-
-  // &:focus
-  //   border: 1px solid #66afe9
-    // outline: 0
-    // box-shadow: inset 0 1px 1px rgba(0,0,0,.075),0 0 3px rgba(#66afe9,.6)
-
-input[type="text"]
-input[type="search"]
-input[type="email"]
-input[type="number"]
-input[type="password"]
-select
-  @extend $input
-  height: em($base-line-height / 2 * 3)
-
-  // &:read-only,
-  // &:disabled
-  //  background lighten(#000, 95%)
-
-select
-  width: auto
-
-::placeholder
-  color: lighten(#000, 55%)
-  font-size: em($base-font-size - 2)
