--- site/src/styles/partials/links.styl
@@ -1,11 +1,0 @@
-a
-  color: #FFF
-
-  &:visited
-    color: #FFF
-
-  &:hover
-    color: #FFFF32
-
-  &:active
-    color: hue(#FFFF32, 10%)
