+++ site/src/styles/partials/paragraph.scss
@@ -1,0 +1,7 @@
+p {
+  margin-top: 1.5em;
+  margin-bottom: 1.5em;
+  font-size: 1em;
+  hyphens: auto;
+  line-height: 1.5em;
+}
