--- site/src/styles/partials/paragraph.styl
@@ -1,5 +1,0 @@
-p
-  adjust-font-size-to($base-font-size)
-  rhythm-margins(1, 1, $base-font-size)
-  hyphens: auto
-  // max-width ($base-font-size * 30)
