+++ site/src/styles/partials/pre.scss
@@ -1,0 +1,4 @@
+pre {
+  overflow: auto;
+  font-feature-settings: 'tnum';
+}
