--- site/src/styles/settings/_variables.styl
@@ -1,37 +1,0 @@
-// Color
-$primary-color = #217bb8
-
-$bg-color = $primary-color
-$selection-color = #FFF
-$selection-bg-color = $primary-color
-
-// Typography
-$font-family-mono = "Courier New", Courier, monospace
-$font-family-base = "Inter", Helvetica, Arial, sans-serif
-
-$base-font-size = 16px
-$base-line-height = 24px
-$base-text-color = #FFF
-
-$font-size-h1 = 36px
-$font-size-h2 = 32px
-$font-size-h3 = 30px
-$font-size-h4 = 26px
-$font-size-h5 = 24px
-$font-size-h6 = 20px
-
-
-// Vertical Rhythmn
-$debug-vertical-rhythm = false
-$debug-color-line = #FFF
-
-
-// Media queries
-
-// The EMs have it: Proportional Media Queries FTW!
-// http://blog.cloudfour.com/the-ems-have-it-proportional-media-queries-ftw/
-// 7 Habits of Highly Effective Media Queries
-// http://bradfrost.com/blog/post/7-habits-of-highly-effective-media-queries/
-$phone = 480px
-$tablet = 768px
-$desktop = 992px
